requirements.txt
certifi==2019.3.9
chardet==3.0.4
Deprecated==1.2.5
idna==2.8
PyGithub==1.43.5
PyJWT==1.7.1
requests==2.21.0
urllib3==1.24.1
wrapt==1.11.1
